Split index argument parsing
